Update Hall constructors
The first now takes a const Player& as opposed to a Player&.
The second now calls the first to create a new Hall with the same seed and Player. May want to change this to create a new Player, if Player doesn't exist independent of Hall - depends on later code.