Start functionality to add Menu option
Effectively a glorified version of addButton. Note that creating and adding a LinkedButton outside the Menu (as seen in Main) works, but trying to run addMenuOption() creates a seg fault on the line sf::Mouse().getPosition(window) in touchingMouse() due to the window variable referencing 0x0. Still tracking down the source of this...